MF-031 Refactor move findAllRoutesWithFare
