+++ src/data/MetroGraph.ts
@@ -1,6 +1,5 @@
 import { MRT_BLUE_STATION_ID, BTS_SILOM_STATION_ID, Graph, BTS_SUKHUMVIT_STATION_ID, ARL_STATION_ID, BRT_STATION_ID, Intersection, METRO_STATION_ID, BTS_GOLD_ID } from "../types";
-import { MRT_BLUE_LINE, BTS_SILOM_LINE, BTS_SUKHUMVIT_LINE, ARL_LINE, BRT_LINE } from ".";
-import { BTS_GOLD_LINE } from "./BtsGoldLine";
+import { BTS_GOLD_LINE, ARL_LINE, BRT_LINE, BTS_SILOM_LINE, BTS_SUKHUMVIT_LINE, MRT_BLUE_LINE } from ".";
 
 export const ALL_INTERSECTIONS: Intersection[] = [
     [MRT_BLUE_STATION_ID.SILOM, BTS_SILOM_STATION_ID.SALA_DAENG],
